make python 3.12 experimental for ga
