+++ .github/workflows/auto_deploy_on_release.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v2
         with:
-          python-version: "3.12"
+          python-version: "3.11"
 
       - name: Pypa build
         run: |
@@ -40,7 +40,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v2
         with:
-          python-version: "3.12"
+          python-version: "3.11"
 
       - name: Upgrade pip
         run: |
