+++ .github/workflows/manual_docs.yml
@@ -14,7 +14,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v2
         with:
-          python-version: "3.12"
+          python-version: "3.11"
 
       - name: Upgrade pip
         run: |
